+++ src/main/java/com/unrealdinnerbone/yaum/api/item/IYaumItem.java
@@ -1,13 +1,13 @@
 package com.unrealdinnerbone.yaum.api.item;
 
 import com.unrealdinnerbone.yaum.api.register.YaumRegistry;
-import com.unrealdinnerbone.yaum.yaum;
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
 import net.minecraft.item.Item;
-import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.client.event.ModelRegistryEvent;
 import net.minecraftforge.client.model.ModelLoader;
 import net.minecraftforge.event.RegistryEvent;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
 
 public interface IYaumItem {
     String getItemName();
@@ -21,6 +21,7 @@ public interface IYaumItem {
         event.getRegistry().register(getItem());
     }
 
+    @SideOnly(Side.CLIENT)
     default void renderItem(ModelRegistryEvent event, YaumRegistry registry) {
         registry.getLogHelper().debug("Rendering Item... " + getItemName());
         ModelLoader.setCustomModelResourceLocation(getItem(), 0, new ModelResourceLocation(getItem().getRegistryName(), "inventory"));