+++ src/main/java/com/unrealdinnerbone/yaum/client/event/rendering/EventPlayerRender.java
@@ -1,16 +1,21 @@
 package com.unrealdinnerbone.yaum.client.event.rendering;
 
-import com.unrealdinnerbone.yaum.Yaum;
 import com.unrealdinnerbone.yaum.config.YaumConfiguration;
 import com.unrealdinnerbone.yaum.libs.Reference;
 import com.unrealdinnerbone.yaum.libs.helpers.GuiHelper;
+import com.unrealdinnerbone.yaum.yaum;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraftforge.client.event.GuiScreenEvent;
 import net.minecraftforge.client.event.RenderGameOverlayEvent;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.eventhandler.EventPriority;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Mod.EventBusSubscriber(modid = Reference.MOD_ID, value = Side.CLIENT)
 @SideOnly(Side.CLIENT)
@@ -27,7 +32,7 @@ public class EventRenderGameOverlay
             }
 
         } catch (Exception e) {
-            Yaum.getRegistry().getLogHelper().error("And error happened while trying to render and GUI Server Overlay : " + e.getMessage());
+            yaum.getRegistry().getLogHelper().error("And error happened while trying to render and GUI Server Overlay : " + e.getMessage());
         }
 
     }