+++ src/main/java/com/unrealdinnerbone/yaum/config/YaumConfiguration.java
@@ -8,35 +8,58 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 
 @Mod.EventBusSubscriber(modid = Reference.MOD_ID)
-public class YaumConfiguration
-{
-    @Config(modid = Reference.MOD_ID, name = Reference.MOD_ID + "/client", category = "supporters")
+public class YaumConfiguration {
+
+    @Config(modid = Reference.MOD_ID, name = Reference.MOD_ID + "/client")
     public static class ClientConfig {
 
-        @Config.RequiresWorldRestart
-        @Config.Comment("Do supports capes get rendered?")
-        public static boolean capesEnabled = true;
+        public static class Supports {
 
-        @Config.RequiresWorldRestart
-        @Config.Comment("Do supports elytra get changed?")
-        public static boolean elytraEnabled = true;
+            @Config.RequiresWorldRestart
+            @Config.Comment("Do supports capes get rendered?")
+            public static boolean capesEnabled = true;
+
+
+            @Config.RequiresWorldRestart
+            @Config.Comment("Do supports elytra get changed?")
+            public static boolean elytraEnabled = true;
+
+
+        }
+
+        public static class Tweaks {
+
+            @Config.Comment("Remove the recipe book")
+            public static boolean disableRecipeBook = true;
+
+            @Config.RequiresMcRestart
+            @Config.Comment("Change the texture of iron nugget")
+            public static boolean changeIronNuggetTexture = true;
+
+        }
 
     }
 
     @Config(modid = Reference.MOD_ID, name = Reference.MOD_ID + "/general", category = "registering")
     public static class GeneralConfig {
 
-        @Config.RequiresMcRestart
-        @Config.Comment("Register the test item?")
-        public static boolean testBlockEnabled = true;
+        public static class Debug {
 
-        @Config.RequiresMcRestart
-        @Config.Comment("Register the test block?")
-        public static boolean testItemEnabled = true;
+            @Config.RequiresMcRestart
+            @Config.Comment("Register the test item?")
+            public static boolean testBlockEnabled = true;
 
+            @Config.RequiresMcRestart
+            @Config.Comment("Register the test block?")
+            public static boolean testItemEnabled = true;
 
-        @Config.Comment("Can Supporters use colors in their chat messages")
-        public static boolean supporterChat = true;
+        }
+
+        public static class General {
+
+            @Config.Comment("Can Supporters use colors in their chat messages")
+            public static boolean supporterChat = true;
+        }
 
     }
 
@@ -47,6 +70,5 @@ public class YaumConfiguration
             ConfigManager.sync(Reference.MOD_ID, Config.Type.INSTANCE);
         }
     }
-
 
 }