+++ src/main/java/com/unrealdinnerbone/yaum/events/registeies/EventBlockRegister.java
@@ -4,16 +4,20 @@ import com.unrealdinnerbone.yaum.api.register.YaumRegistry;
 import com.unrealdinnerbone.yaum.libs.Reference;
 import net.minecraft.block.Block;
 import net.minecraftforge.event.RegistryEvent;
+import net.minecraftforge.fml.common.FMLCommonHandler;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.ModContainer;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 
 @Mod.EventBusSubscriber(modid = Reference.MOD_ID)
 public class EventBlockRegister
 {
     @SubscribeEvent
-    public void registerBlocks(RegistryEvent.Register<Block> event) {
+    public static void registerBlocks(RegistryEvent.Register<Block> event) {
         for(YaumRegistry yaumRegistry: YaumRegistry.getRegistries().values()) {
             if (yaumRegistry.getRegisteredBlocks() != null && yaumRegistry.getRegisteredBlocks().size() > 0) {
+                ModContainer container = FMLCommonHandler.instance().findContainerFor(yaumRegistry.getModID());
+                event.setModContainer(container);
                 yaumRegistry.getLogHelper().info("Starting Block Regeneration for " + yaumRegistry.getRegisteredBlocks().size()  + " block(s)");
                 yaumRegistry.getRegisteredBlocks().forEach(iYaumBlock -> iYaumBlock.registerBlock(event, yaumRegistry));
                 yaumRegistry.getLogHelper().info("Finished Block Regeneration for " + yaumRegistry.getRegisteredBlocks().size()  + " block(s)");