+++ src/main/java/com/unrealdinnerbone/yaum/libs/Reference.java
@@ -6,8 +6,8 @@ public class Reference
 {
   public static final String MOD_ID = "yaum";
   public static final String MOD_NAME = "YAUM - (Yet Another util Mod)";
-  public static final String VERSION = "1.3.0";
-  public static final String MC_VERSION = "1.12";
+  public static final String VERSION = "1.4.0";
+  public static final String MC_VERSION = "1.12.2";
   public static final String SERVER_SIDE = "com.unrealdinnerbone.yaum.proxy.ServerProxy";
   public static final String CLIENT_SIDE = "com.unrealdinnerbone.yaum.proxy.ClientProxy";
   public static final String GUI_FACTORY_CLASS = "com.unrealdinnerbone.yaum.client.gui.ModConfigGUI";