+++ src/main/java/com/unrealdinnerbone/yaum/libs/YaumCreativeTab.java
@@ -1,7 +1,7 @@
 package com.unrealdinnerbone.yaum.libs;
 
+import com.unrealdinnerbone.yaum.Yaum;
 import com.unrealdinnerbone.yaum.api.util.LangHelper;
-import com.unrealdinnerbone.yaum.yaum;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.init.Items;
 import net.minecraft.item.ItemStack;
@@ -9,7 +9,7 @@ import net.minecraft.item.ItemStack;
 public class YaumCreativeTab extends CreativeTabs {
 
     public YaumCreativeTab(String unlocalLisedName) {
-        super(yaum.getRegistry().getLangHelper().translateMessage(LangHelper.Type.TAB, unlocalLisedName));
+        super(Yaum.getRegistry().getLangHelper().translateMessage(LangHelper.Type.TAB, unlocalLisedName));
     }
 
     @Override