+++ src/main/java/com/unrealdinnerbone/yaum/libs/utils/DateUtils.java
@@ -1,23 +1,22 @@
 package com.unrealdinnerbone.yaum.libs.utils;
 
-import com.unrealdinnerbone.yaum.yaum;
+import com.unrealdinnerbone.yaum.Yaum;
 
 import java.util.Calendar;
 
 public class DateUtils {
 
     public static void checkDates() {
-
         if(isToday(1, 1) || isToday(12, 31)) {
-            yaum.getRegistry().getLogHelper().info("Happy New Year");
+            Yaum.getRegistry().getLogHelper().info("Happy New Year");
         }else if(isToday(4, 1))        {
-            yaum.getRegistry().getLogHelper().info("Happy April Fools Day");
+            Yaum.getRegistry().getLogHelper().info("Happy April Fools Day");
         }else if(isToday(10, 31)) {
-            yaum.getRegistry().getLogHelper().info("Happy Halloween");
+            Yaum.getRegistry().getLogHelper().info("Happy Halloween");
         }else if(isToday(12, 24) || isToday(12, 25)) {
-            yaum.getRegistry().getLogHelper().info("Merry Christmas");
+            Yaum.getRegistry().getLogHelper().info("Merry Christmas");
         }else {
-            yaum.getRegistry().getLogHelper().info(":( Today is not a Holiday");
+            Yaum.getRegistry().getLogHelper().info(":( Today is not a Holiday");
         }
 
     }