+++ src/main/java/com/unrealdinnerbone/yaum/perks/Supporters.java
@@ -1,0 +1,4 @@
+package com.unrealdinnerbone.yaum.perks;
+
+public class Supporters {
+}