+++ src/main/java/com/unrealdinnerbone/yaum/proxy/ClientProxy.java
@@ -1,5 +1,9 @@
 package com.unrealdinnerbone.yaum.proxy;
 
+import com.unrealdinnerbone.yaum.client.render.PlayerRenderLayer;
+import com.unrealdinnerbone.yaum.libs.helpers.DownloadHelper;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.renderer.entity.RenderPlayer;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
@@ -22,6 +26,10 @@ public class ClientProxy extends CommonProxy {
     @Override
     public void onPostInt(FMLPostInitializationEvent event) {
         super.onPostInt(event);
+        PlayerRenderLayer layer = new PlayerRenderLayer();
+        for (RenderPlayer playerRender : Minecraft.getMinecraft().getRenderManager().getSkinMap().values()) {
+            playerRender.addLayer(layer);
+        }
     }
 
 }